requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==20.2.2
Automat==20.2.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2019.6.16
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
flake8==3.7.8
flake8-docstrings==1.3.1
flake8-polyfill==1.0.2
Flask==1.1.1
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
gunicorn==19.9.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.6.2
numpy==1.17.4
oauthlib==3.1.0
packaging==19.1
paramiko==2.7.1
pluggy==0.13.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==4.0.1
pyflakes==2.1.1
PyHamcrest==2.0.0
pymongo==3.9.0
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.2
pyrsistent==0.15.7
pytest==5.1.2
python-dotenv==0.10.3
python-monkey-business==1.0.0
pytz==2019.3
PyYAML==5.3
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
snowballstemmer==1.9.0
sqlparse==0.3.0
texttable==1.6.2
Twisted==19.10.0
TwitterAPI==2.5.10
txaio==20.1.1
urllib3==1.25.3
utils==0.9.0
virtualenv==16.7.3
wcwidth==0.1.7
websocket-client==0.57.0
Werkzeug==0.15.6
WTForms==2.2.1
zipp==0.6.0
zope.interface==4.7.1